Integrated database into log
security_LogsDB is fully intergared into the security api.
A script has been made to check the uploaded file and the connection of the database.
The AI api does not work yet due to credits, the standard mode works and the database is connected.
Start script cant be used without the password.